Use float16 for T4; remove MiniCPM (no T4-suitable real model)
- All VLMs + judge load in float16 (T4/Turing has no bf16 tensor cores).
- Drop MiniCPM from the registry, runner factory, and delete its adapter:
  the configured openbmb/MiniCPM-V-4_6 does not exist on HuggingFace and the
  real small variant (MiniCPM-V-2) uses an incompatible legacy chat API.

+++ judge/qwen_judge.py
@@ -52,13 +52,13 @@ class QwenJudge:
         if self.load_in_4bit:
             kwargs["quantization_config"] = BitsAndBytesConfig(
                 load_in_4bit=True,
-                bnb_4bit_compute_dtype=torch.bfloat16,
+                bnb_4bit_compute_dtype=torch.float16,  # T4 (Turing) has no bf16 tensor cores
                 bnb_4bit_quant_type="nf4",
                 bnb_4bit_use_double_quant=True,
             )
             kwargs["device_map"] = "auto"
         else:
-            kwargs["torch_dtype"] = torch.bfloat16
+            kwargs["torch_dtype"] = torch.float16
         self.model = AutoModelForCausalLM.from_pretrained(self.model_id, **kwargs)
         if not self.load_in_4bit:
             self.model = self.model.to(self.device)
